# Replace Quadtree with more accurate implementation