Supersedes #925 (ported over tests)

## Implementation

After benchmark testing with [Orthogonal Arrays](https://en.wikipedia.org/wiki/Orthogonal_array), the minimum node geometric size and max items per node parameters were tweaked for better performance.

Additionally, an overlap parameter was added. This scales the subgrids by the provided factor, such that each subgrid overlaps its neighbors. This helps prevent hotspots on cell borders from piling up in the parent nodes.

This new implementation persists the existing behavior that two entities whose borders perfectly touch are **not** considered colliding. If that is not the desired behavior we may wish to address that separately.

This also switches to using an interface in the method signatures instead of the class name itself, partially because that made it easier for me to flip back and forth between testing and partially so that if someone else ever wants to burn this quadtree with fire and replace it with their implementation that's even better, it will be easier for them as well :)

## Benchmarks vs previous implementation
### Insert
I would have expected the performance here to be similar, so it's interesting the new quadtree is so much faster.
| Method         | MaxItemSize | InsertedItems | Mean       | Error     | StdDev    | Ratio |
|--------------- |------------ |-------------- |-----------:|----------:|----------:|------:|
| InsertQuadtree | 15          | 5000          | 839.493 us | 8.9584 us | 8.3797 us | 1.000 |
| InsertNewTree   | 15          | 5000          |   3.346 us | 0.0241 us | 0.0226 us | 0.004 |

### Update
I'm not quite sure why the old quadtree updates were so slow. This benchmark also isn't actually recording the performance of a single update, but rather multiple.
| Method         | TotalItems | MovingItems | Mean         | Error      | StdDev     | Ratio |
|--------------- |----------- |------------ |-------------:|-----------:|-----------:|------:|
| UpdateQuadtree | 15000      | 10000       | 3,205.934 ms | 11.2854 ms | 10.0043 ms | 1.000 |
| UpdateNewTree   | 15000      | 10000       |     4.007 ms |  0.0544 ms |  0.0455 ms | 0.001 |

### Query
Here it's no surprise that the new quadtree is slower - many items were missed in the previous implementation, so it's not really an apples-to-apples comparison.
| Method            | ItemCount | ItemSize | AreaSideLength | Mean       | Error    | StdDev  | Ratio | RatioSD |
|------------------ |---------- |--------- |--------------- |-----------:|---------:|--------:|------:|--------:|
| QueryAreaQuadtree | 3000      | 10       | 100            |   406.1 ns |  2.62 ns | 2.05 ns |  1.00 |    0.00 |
| QueryAreaNewTree   | 3000      | 10       | 100            | 1,634.2 ns | 10.37 ns | 9.19 ns |  4.02 |    0.04 |

## In-game comparison
### Old
![](https://media.giphy.com/media/v1.Y2lkPTc5MGI3NjExbTZnODM5NTlneGltaDN1am92Ympjc3o0cTg5b24yNGNpeGRkZ2MzdiZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/qBoyiBiSJgOVw35yFN/giphy.gif)
### New
![](https://media.giphy.com/media/v1.Y2lkPTc5MGI3NjExYXhhb3F2aGY2aXh3cTZybTB1bDIzOHdqcjR2MmYwdWZvcHl0NWI4ZyZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/NcgrYoDIDyTpRhtN5i/giphy.gif)