@tgjones 2024-08-01
This is awesome!

Often when the benchmark results are wildly different from expectations, it's worth double checking that the benchmarks actually measure what you think they do 😄 But assuming they do, then this looks like a very nice speedup on insert and update, with an expected slowdown for query since it's returning more data.