@charliefoxtwo 2024-08-01
More benchmarks here couldn't hurt (remove isn't benchmarked at all currently), that's for sure. Since the new quadtree has an identical public signature to the old one, they used the exact same code in the benchmark methods, so I'm pretty comfortable saying they _should have_ been testing the same exact same thing.

I bet this could be further optimized for even better performance as well - but I'm not super experienced in that area.